+++ contours/contours_modal.py
@@ -29,12 +29,12 @@ import math
 import os
 
 
-from .modaloperator import ModalOperator
-from . import key_maps
-from .lib import common_utilities
-from .lib.common_utilities import bversion, get_object_length_scale, dprint, profiler, frange, selection_mouse, showErrorMessage
+from ..modaloperator import ModalOperator
+from .. import key_maps
+from ..lib import common_utilities
+from ..lib.common_utilities import bversion, get_object_length_scale, dprint, profiler, frange, selection_mouse, showErrorMessage
 from .contour_classes import Contours
-from .lib.common_utilities import showErrorMessage
+
 
 class  CGC_Contours(ModalOperator):
     '''Draw Strokes Perpindicular to Cylindrical Forms to Retopologize Them'''
